refactor: Move PageMeta into Page component file
